♻️Moved inferr_file_format and not_implemented_to_value_error to new module
glotaran/plugin_system/io_plugin_utils.py
