+++ glotaran/plugin_system/base_registry.py
@@ -6,14 +6,10 @@ This is to prevent issues with circular imports.
 """
 from __future__ import annotations
 
-import functools
-import os
 from importlib import metadata
 from typing import TYPE_CHECKING
 from typing import Any
-from typing import Callable
 from typing import TypeVar
-from typing import cast
 from warnings import warn
 
 if TYPE_CHECKING:
@@ -28,9 +24,6 @@ if TYPE_CHECKING:
     _PluginInstantiableType = TypeVar(
         "_PluginInstantiableType", DataIoInterface, ProjectIoInterface
     )
-
-
-F = TypeVar("F", bound=Callable[..., Any])  # decorated function
 
 
 class __PluginRegistry:
@@ -303,63 +296,3 @@ def get_plugin_from_registry(
         raise ValueError(not_found_error_message)
     else:
         return plugin_registry[plugin_register_key]
-
-
-def inferr_file_format(file_path: str | os.PathLike[str]) -> str:
-    """Inferr format of a file if it exists.
-
-    Parameters
-    ----------
-    file_path : str
-        Path/str to the file.
-
-    Returns
-    -------
-    str
-        File extension without the leading dot.
-
-    Raises
-    ------
-    ValueError
-        If file doesn't exists.
-    ValueError
-        If file has no extension.
-    """
-    if not os.path.isfile(file_path):
-        raise ValueError(f"There is no file {file_path!r}.")
-
-    _, file_format = os.path.splitext(file_path)
-    if file_format == "":
-        raise ValueError(
-            f"Cannot determine format of file {file_path!r}, please provide an explicit format."
-        )
-    else:
-        return file_format.lstrip(".")
-
-
-def not_implemented_to_value_error(func: F) -> F:
-    """Decorate a function to raise ValueError instead of NotImplementedError.
-
-    This decorator is supposed to be used on functions which call functions
-    that might raise a NotImplementedError, but raise ValueError instead with
-    the same error text.
-
-    Parameters
-    ----------
-    func : Callable[..., RT]
-        Function to be decorated.
-
-    Returns
-    -------
-    Callable[..., RT]
-        Wrapped function.
-    """
-
-    @functools.wraps(func)
-    def wrapper(*args: Any, **kwargs: Any) -> Any:
-        try:
-            return func(*args, **kwargs)
-        except NotImplementedError as error:
-            raise ValueError(error.args)
-
-    return cast(F, wrapper)
